+++ src/main/java/com/nemonotfound/nemoswoodcutter/NemosWoodcutterClient.java
@@ -1,15 +1,12 @@
-package com.nemonotfound;
+package com.nemonotfound.nemoswoodcutter;
 
-import com.nemonotfound.block.ModBlocks;
-import com.nemonotfound.screen.WoodcutterScreen;
+import com.nemonotfound.nemoswoodcutter.block.ModBlocks;
+import com.nemonotfound.nemoswoodcutter.screen.WoodcutterScreen;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.render.RenderLayer;
 
-@Environment(EnvType.CLIENT)
 public class NemosWoodcutterClient implements ClientModInitializer {
 
     @Override