+++ Changelog.md
@@ -2,7 +2,9 @@
 
 ## Changes
 - Updated Fabric Loader to 0.15.10
-- 
+
+## Fixes
+- Game crashes when using Nemo's Mossy Blocks with Nemo's Woodcutter
 
 ### Downloads
 [CurseForge](https://curseforge.com/minecraft/mc-mods/nemos-woodcutter) <br>